修改 使用uint作为Id类型 修改 Response的StatusCode类型改为int

+++ src/dto/user_register_response.go
@@ -2,6 +2,6 @@ package dto
 
 type UserRegisterResponse struct {
 	*Response
-	UserId int64  `json:"user_id"`
+	UserId uint   `json:"user_id"`
 	Token  string `json:"token"`
 }
